Restructure packages to void camel case in files names
