+++ pkg/controller/workspace/component/dockerimage.go
@@ -28,7 +28,7 @@ import (
 	. "github.com/che-incubator/che-workspace-crd-operator/pkg/controller/workspace/model"
 )
 
-func setupDockerImageComponent(names WorkspaceProperties, commands []workspaceApi.CommandSpec, component *workspaceApi.ComponentSpec) (*ComponentInstanceStatus, error) {
+func setupDockerimageComponent(names WorkspaceProperties, commands []workspaceApi.CommandSpec, component *workspaceApi.ComponentSpec) (*ComponentInstanceStatus, error) {
 	componentInstanceStatus := &ComponentInstanceStatus{
 		Machines:                   map[string]MachineDescription{},
 		Endpoints:                  []workspaceApi.Endpoint{},
